refactor(skills): colocate the simple creation page and form - Put SimpleCreateSkillPage in its own page directory - Keep form state, validation, and submission in its dedicated component
