+++ src/pages/skills/simple-create-skill/components/skill-create-form.tsx
@@ -1,0 +1,110 @@
+import { type FormEvent, useState } from "react";
+import {
+	Button,
+	Input,
+	Label,
+	TextArea,
+	TextField,
+	Toast,
+} from "@heroui/react";
+import { useTranslation } from "react-i18next";
+import { useNavigate } from "react-router";
+import { handleError } from "@/utils/error";
+import { useCreatePlatformSkill } from "@/queries/skill";
+
+/** Keeps simple creation fields and submission feedback together within the dedicated page. */
+const SkillCreateForm = () => {
+	const { t } = useTranslation();
+	const navigate = useNavigate();
+	const createSkillMutation = useCreatePlatformSkill();
+	const [displayName, setDisplayName] = useState("");
+	const [description, setDescription] = useState("");
+	const [content, setContent] = useState("");
+	const trimmedDisplayName = displayName.trim();
+	const isDisplayNameValid = /^(?=.*[A-Za-z])[A-Za-z0-9 _-]+$/.test(
+		trimmedDisplayName,
+	);
+	const canCreate = Boolean(
+		isDisplayNameValid && description.trim() && content.trim(),
+	);
+
+	const submit = async (event: FormEvent<HTMLFormElement>) => {
+		event.preventDefault();
+		if (!canCreate || createSkillMutation.isPending) return;
+		try {
+			await createSkillMutation.mutateAsync({
+				displayName: trimmedDisplayName,
+				description: description.trim(),
+				content: content.trim(),
+			});
+			Toast.toast.success(
+				t("skills.create.success", { skill: trimmedDisplayName }),
+			);
+			navigate("/skills");
+		} catch (error) {
+			handleError(
+				error,
+				"Platform Skill creation failed",
+				true,
+				t("skills.create.failed"),
+			);
+		}
+	};
+
+	return (
+		<form
+			className="mt-xl flex flex-col gap-lg rounded-lg border border-hairline bg-surface-card p-lg sm:p-xl"
+			onSubmit={submit}
+		>
+			<TextField className="flex flex-col gap-xs text-body-sm font-medium text-ink">
+				<Label>{t("skills.create.nameLabel")}</Label>
+				<Input
+					className="rounded-md border border-hairline bg-canvas px-md py-sm font-normal outline-none focus:border-hairline-strong focus:ring-2 focus:ring-focus-ring"
+					maxLength={120}
+					onChange={(event) => setDisplayName(event.target.value)}
+					value={displayName}
+				/>
+				{trimmedDisplayName && !isDisplayNameValid ? (
+					<p className="text-caption-sm font-normal text-terminal-red">
+						{t("skills.create.nameInvalid")}
+					</p>
+				) : null}
+			</TextField>
+			<TextField className="flex flex-col gap-xs text-body-sm font-medium text-ink">
+				<Label>{t("skills.create.descriptionLabel")}</Label>
+				<TextArea
+					className="min-h-24 resize-y rounded-md border border-hairline bg-canvas px-md py-sm font-normal outline-none focus:border-hairline-strong focus:ring-2 focus:ring-focus-ring"
+					onChange={(event) => setDescription(event.target.value)}
+					value={description}
+				/>
+			</TextField>
+			<TextField className="flex flex-col gap-xs text-body-sm font-medium text-ink">
+				<Label>{t("skills.create.contentLabel")}</Label>
+				<TextArea
+					className="min-h-72 resize-y rounded-md border border-hairline bg-canvas px-md py-sm font-mono text-body-sm font-normal leading-6 outline-none focus:border-hairline-strong focus:ring-2 focus:ring-focus-ring"
+					onChange={(event) => setContent(event.target.value)}
+					value={content}
+				/>
+			</TextField>
+			{createSkillMutation.error ? (
+				<p className="text-body-sm text-terminal-red" role="alert">
+					{t("skills.create.failed")}
+				</p>
+			) : null}
+			<div className="flex justify-end gap-sm border-t border-hairline pt-lg">
+				<Button onPress={() => navigate("/skills")} variant="tertiary">
+					{t("common.cancel")}
+				</Button>
+				<Button
+					isDisabled={!canCreate || createSkillMutation.isPending}
+					type="submit"
+					variant="primary"
+				>
+					{t("skills.create.submit")}
+				</Button>
+			</div>
+		</form>
+	);
+};
+
+export { SkillCreateForm };
